\name{uniPlot-internal}
\alias{theme_uniPlot}
\alias{theme_text2}
\alias{par_uniPlot}
\alias{theme_lattice_par_uniPlot}
\alias{theme_lattice_options_uniPlot}
\alias{init}
\alias{synchron}
\title{Internal uniPlot objects}
\description{Internal uniPlot objects.}
\details{These are not to be called by the user.}
\keyword{internal}